refactor(mini cart): Clean up after merge commit
