+++ packages/auth-sdk/package.json
@@ -9,7 +9,9 @@
       "@ifixit/helpers": "workspace:*"
    },
    "peerDependencies": {
-      "react-query": "3.x"
+      "react-query": "3.x",
+      "next": "*",
+      "cookie": "0.5.0"
    },
    "devDependencies": {
       "@types/react": "17.0.37",
